requirements.txt
discord.py
wheel
jishaku
aioredis